ua_geotransformer_repro_20251001/requirements.txt
numpy
pandas
pyyaml
rasterio
scikit-learn
scipy
shap
torch
tqdm
xgboost